Refactoring BrandService and start using NgRx Store

+++ project.frontend/ClientApp/src/app/brands/show-brand/show-brand.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs';
-import { Brand } from '../../dto/brand.model';
+import { Brand } from '../../dto/Brand.interface';
 import { BrandService } from '../../services/brand.service';
 
 @Component({
@@ -21,31 +21,14 @@ export class ShowBrandComponent implements OnInit {
     private brandService: BrandService,
     private router: Router
   ) {
-    /*    this.brandService.GetConf().ReloadConf();*/
-  }
-
-  private TempFunction(): any {
-    //console.log("Valore isLoadingSetting Conf " + this.brandService.GetConf().isLoading)
   }
 
   ngOnInit(): void {
-    window.setInterval(() => this.TempFunction(), 100);
 
     //SUBSCRIBE PER RICHIAMARE L'ID DALL'URL
     this._Activatedroute.paramMap.subscribe(params => {
       this.brandId = Number(params.get("brandId"));
 
-      //this.brandService.GetConf().testBoolean
-      //  .subscribe((response) => {
-      //    console.log("Precall Brandservice");
-      //    this.brandService
-      //      .FindByIdAsync(this.brandId)
-      //      .pipe(
-      //        map((res) => this.actualBrand = res)
-      //      )
-      //      .subscribe((response) => (this.responseReceived(response)))
-      //    console.log("Endcall Brandservice");
-      //  })
       console.log("Precall Brandservice");
       this.brandService
         .FindByIdAsync(this.brandId)
@@ -70,19 +53,21 @@ export class ShowBrandComponent implements OnInit {
   public DeleteAsync(id: number) {
     this.loading = true;
     this.errorMessage = "";
-    this.brandService.DeleteAsync(id)
-      .subscribe(
-        (response) => {                           //next() callback
-          this.router.navigate(['/brands']);
-        },
-        (error) => {                              //error() callback
-          console.error('Request failed with error')
-          this.errorMessage = error;
-          this.loading = false;
-        },
-        () => {                                   //complete() callback
-          console.info('Request completed')      //This is actually not needed 
-          this.loading = false;
-        })
+    if (confirm("Are you sure do you want to delete this Brand?")) {
+      this.brandService.DeleteAsync(id)
+        .subscribe(
+          (response) => {                           //next() callback
+            this.router.navigate(['/brands']);
+          },
+          (error) => {                              //error() callback
+            console.error('Request failed with error')
+            this.errorMessage = error;
+            this.loading = false;
+          },
+          () => {                                   //complete() callback
+            console.info('Request completed')      //This is actually not needed 
+            this.loading = false;
+          })
+    }
   }
 }
